requirements.txt
Django
Pillow
django-crispy-forms
crispy-bootstrap5
psycopg2-binary